sketch/pyproject.toml
[project]
name = "sketch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-discoveryengine>=0.13.12",
]
